feat: Enhance scoring system with komi and seki detection
- Updated displayScore method to support fractional scores and declare winners.
- Added komi information display in JavaFXGameClient.
- Introduced komi handling in Board class, affecting score calculations.
- Implemented seki detection logic to identify mutual life situations on the board.
- Enhanced AsciiBoardRenderer and ConsoleGameUI to reflect seki and komi in outputs.
- Added comprehensive tests for komi and seki functionalities, ensuring correct behavior in various scenarios.
- Updated existing tests to accommodate changes in score representation and game rules.

+++ src/main/java/com/example/goboard/view/GameUI.java
@@ -77,8 +77,9 @@ public interface GameUI {
 
     /**
      * Display the currently calculated score.
+     * Supports fractional scores due to komi.
      */
-    void displayScore(int blackScore, int whiteScore);
+    void displayScore(double blackScore, double whiteScore);
 
     /**
      * Ask the player to confirm the final score.
